requirements.txt
beautifulsoup4==4.9.3
blinker==1.4
bs4==0.0.1
cached-property==1.5.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
commonmark==0.9.1
dictdiffer==0.8.1
Flask==1.1.2
Flask-Mail==0.9.1
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
git+https://github.com/c0j0s/notion-py@master#egg=notion
python-dotenv==0.17.0
python-slugify==4.0.1
pytz==2021.1
requests==2.25.1
soupsieve==2.2.1
termcolor==1.1.0
text-unidecode==1.3
tzlocal==2.1
urllib3==1.26.4
Werkzeug==1.0.1
